feat(claim): setup claim page
