+++ components/drink/Badge.tsx
@@ -1,9 +1,9 @@
 import { Center, Image } from "@chakra-ui/react";
 
-const Badge = ({ badge }) => {
+const Badge = ({ path }) => {
   return (
     <Center w={"100%"} py={"5em"}>
-      <Image src={badge} boxSize={"sm"} alt="drink badge" />
+      <Image src={path} boxSize={"sm"} alt="drink badge" />
     </Center>
   );
 };
